moved prepare to its own file
